+++ packages/core-js/actual/iterator/drop.js
@@ -1,8 +1,6 @@
 'use strict';
-require('../../modules/es.object.to-string');
+var parent = require('../../stable/iterator/drop');
 require('../../modules/esnext.iterator.constructor');
 require('../../modules/esnext.iterator.drop');
 
-var entryUnbind = require('../../internals/entry-unbind');
-
-module.exports = entryUnbind('Iterator', 'drop');
+module.exports = parent;