+++ packages/core-js/actual/iterator/every.js
@@ -1,8 +1,6 @@
 'use strict';
-require('../../modules/es.object.to-string');
+var parent = require('../../stable/iterator/every');
 require('../../modules/esnext.iterator.constructor');
 require('../../modules/esnext.iterator.every');
 
-var entryUnbind = require('../../internals/entry-unbind');
-
-module.exports = entryUnbind('Iterator', 'every');
+module.exports = parent;