+++ packages/core-js/actual/iterator/filter.js
@@ -1,8 +1,6 @@
 'use strict';
-require('../../modules/es.object.to-string');
+var parent = require('../../stable/iterator/filter');
 require('../../modules/esnext.iterator.constructor');
 require('../../modules/esnext.iterator.filter');
 
-var entryUnbind = require('../../internals/entry-unbind');
-
-module.exports = entryUnbind('Iterator', 'filter');
+module.exports = parent;