+++ packages/core-js/actual/iterator/flat-map.js
@@ -1,8 +1,6 @@
 'use strict';
-require('../../modules/es.object.to-string');
+var parent = require('../../stable/iterator/flat-map');
 require('../../modules/esnext.iterator.constructor');
 require('../../modules/esnext.iterator.flat-map');
 
-var entryUnbind = require('../../internals/entry-unbind');
-
-module.exports = entryUnbind('Iterator', 'flatMap');
+module.exports = parent;