+++ packages/core-js/actual/iterator/take.js
@@ -1,8 +1,6 @@
 'use strict';
-require('../../modules/es.object.to-string');
+var parent = require('../../stable/iterator/take');
 require('../../modules/esnext.iterator.constructor');
 require('../../modules/esnext.iterator.take');
 
-var entryUnbind = require('../../internals/entry-unbind');
-
-module.exports = entryUnbind('Iterator', 'take');
+module.exports = parent;