+++ packages/core-js/actual/iterator/to-async.js
@@ -1,6 +1,8 @@
 'use strict';
 require('../../modules/es.object.to-string');
 require('../../modules/es.promise');
+require('../../modules/es.iterator.constructor');
+// TODO: Drop from `core-js@4`
 require('../../modules/esnext.iterator.constructor');
 require('../../modules/esnext.iterator.to-async');